requirements.txt
Faker~=25.2.0
pandas~=2.2.2
numpy~=1.26.4
fastapi~=0.111.0
pydantic~=2.7.1
apache-airflow~=2.9.1